refactor out admin link component
